// payload-types.ts
export interface Page {
  id: number;
  title: string;
  hero: {
    type: 'none' | 'fullScreen' | 'boxed' | 'small';
    heading?: string | null;
    media?: (number | null) | Media;
  };
  layout: (
    | CallToActionBlock
    | StatsBlock
    | MasonryGridBlock
    | TextCalloutBlock
    | FeaturedPropertiesBlock
    | TestimonialBlock
    | AboutMasonryBlock
    | TeamMembersBlock
  )[];
  meta?: {
    title?: string | null;
    /**
     * Maximum upload file size: 12MB. Recommended file size for images is <500KB.
     */
    image?: (number | null) | Media;
    description?: string | null;
  };
  publishedAt?: string | null;
  slug?: string | null;
  slugLock?: boolean | null;
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: number;
  alt?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
  sizes?: {
    thumbnail?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    square?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    small?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    medium?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    large?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    xlarge?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    og?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
  };
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "CallToActionBlock".
 */
export interface CallToActionBlock {
  richText?: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  link: {
    type?: ('reference' | 'custom') | null;
    newTab?: boolean | null;
    reference?: {
      relationTo: 'pages';
      value: number | Page;
    } | null;
    url?: string | null;
    label: string;
  };
  id?: string | null;
  blockName?: string | null;
  blockType: 'cta';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "StatsBlock".
 */
export interface StatsBlock {
  boxOne: {
    title: string;
    subTitle: string;
  };
  boxTwo: {
    title: string;
    subTitle: string;
  };
  boxThree: {
    title: string;
    subTitle: string;
  };
  id?: string | null;
  blockName?: string | null;
  blockType: 'stats';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "MasonryGridBlock".
 */
export interface MasonryGridBlock {
  rowOne: {
    tag?: string | null;
    header: string;
    richText: {
      root: {
        type: string;
        children: {
          type: string;
          version: number;
          [k: string]: unknown;
        }[];
        direction: ('ltr' | 'rtl') | null;
        format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
        indent: number;
        version: number;
      };
      [k: string]: unknown;
    };
    link: {
      type?: ('reference' | 'custom') | null;
      newTab?: boolean | null;
      reference?: {
        relationTo: 'pages';
        value: number | Page;
      } | null;
      url?: string | null;
      label: string;
    };
    media: number | Media;
  };
  rowTwo: {
    tag?: string | null;
    header: string;
    richText: {
      root: {
        type: string;
        children: {
          type: string;
          version: number;
          [k: string]: unknown;
        }[];
        direction: ('ltr' | 'rtl') | null;
        format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
        indent: number;
        version: number;
      };
      [k: string]: unknown;
    };
    link: {
      type?: ('reference' | 'custom') | null;
      newTab?: boolean | null;
      reference?: {
        relationTo: 'pages';
        value: number | Page;
      } | null;
      url?: string | null;
      label: string;
    };
    media: number | Media;
  };
  id?: string | null;
  blockName?: string | null;
  blockType: 'masonryGrid';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "TextCalloutBlock".
 */
export interface TextCalloutBlock {
  header: string;
  content: string;
  id?: string | null;
  blockName?: string | null;
  blockType: 'textCallout';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "FeaturedPropertiesBlock".
 */
export interface FeaturedPropertiesBlock {
  communityOne: number | Community;
  communityTwo: number | Community;
  communityThree: number | Community;
  id?: string | null;
  blockName?: string | null;
  blockType: 'featuredProperties';
}
// ...
export interface TestimonialBlock {
  testimonial: string;
  author: string;
  media: number | Media;
  id?: string | null;
  blockName?: string | null;
  blockType: 'testimonial';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "AboutMasonryBlock".
 */
export interface AboutMasonryBlock {
  heading: string;
  link: {
    type?: ('reference' | 'custom') | null;
    newTab?: boolean | null;
    reference?: {
      relationTo: 'pages';
      value: number | Page;
    } | null;
    url?: string | null;
    label: string;
  };
  headerImage: number | Media;
  values: {
    valueOne: {
      valueName: string;
      valueDescription: string;
      valueIcon: number | Media;
    };
    valueTwo: {
      valueName: string;
      valueDescription: string;
      valueIcon: number | Media;
    };
    valueThree: {
      valueName: string;
      valueDescription: string;
      valueIcon: number | Media;
    };
    valueFour: {
      valueName: string;
      valueDescription: string;
      valueIcon: number | Media;
    };
  };
  images: (number | Media)[];
  id?: string | null;
  blockName?: string | null;
  blockType: 'aboutMasonry';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "TeamMembersBlock".
 */
export interface TeamMembersBlock {
  id?: string | null;
  blockName?: string | null;
  blockType: 'teamMembers';
}